Extracting hooks (#476)
- extracted reusable hooks nd other good stuff to app-ui package
