+++ packages/app-ui/src/hooks/use-local-storage.ts
@@ -1,10 +1,18 @@
+'use client'
 import { type Dispatch, type SetStateAction, useCallback, useEffect, useState } from 'react'
 import { useEventCallback, useEventListener } from 'usehooks-ts'
 
 type SetValue<T> = Dispatch<SetStateAction<T>>
 type isValidFunction<T> = (element?: T) => element is T
 
-function parseJSON<T>(value: string | null, key: string): T | undefined {
+/**
+ * Parses a JSON string from localStorage.
+ *
+ * @param value - The JSON string to parse.
+ * @param key - The key associated with the JSON string, used for error logging.
+ * @returns The parsed value, or undefined if parsing fails.
+ */
+const parseJSON = <T>(value: string | null, key: string): T | undefined => {
   try {
     return typeof value === 'undefined' || value === null ? undefined : JSON.parse(value)
   } catch (error) {
@@ -15,11 +23,19 @@ function parseJSON<T>(value: string | null, key: string): T | undefined {
   }
 }
 
-export function getStorageValue<V>(
+/**
+ * Retrieves a value from localStorage and parses it into a usable format.
+ *
+ * @param key - The localStorage key to retrieve the value from.
+ * @param defaultValue - The default value to return if the key does not exist or parsing fails.
+ * @param isValid - An optional validation function to check the validity of the parsed value.
+ * @returns The stored value if valid, otherwise the default value.
+ */
+export const getStorageValue = <V>(
   key: string,
   defaultValue: unknown,
   isValid?: isValidFunction<V>,
-) {
+) => {
   if (typeof window !== 'undefined') {
     const saved = localStorage.getItem(key)
     const parsed = parseJSON<V>(saved, key)
@@ -34,11 +50,19 @@ export function getStorageValue<V>(
   return defaultValue as V
 }
 
-export function useLocalStorage<T>(
+/**
+ * Custom hook to manage state synchronized with localStorage.
+ *
+ * @param key - The localStorage key to store the value under.
+ * @param initialValue - The initial value to use if the key does not exist in localStorage.
+ * @param isValid - An optional validation function to ensure the stored value is valid.
+ * @returns A tuple with the current value and a setter function to update it.
+ */
+export const useLocalStorage = <T>(
   key: string,
   initialValue: T,
   isValid?: isValidFunction<T>,
-): [T, SetValue<T | null>] {
+): [T, SetValue<T | null>] => {
   const readValue = useCallback((): T => {
     if (typeof window === 'undefined') {
       return initialValue
@@ -62,6 +86,7 @@ export function useLocalStorage<T>(
   }, [initialValue, isValid, key])
 
   const [storedValue, setStoredValue] = useState<T>(readValue)
+
   const setValue: SetValue<T | null> = useEventCallback((value: unknown) => {
     if (typeof window === 'undefined') {
       // eslint-disable-next-line no-console
@@ -90,8 +115,8 @@ export function useLocalStorage<T>(
   }, [])
 
   const handleStorageChange = useCallback(
-    (event: StorageEvent | CustomEvent) => {
-      if ((event as StorageEvent).key && (event as StorageEvent).key !== key) {
+    (ev: StorageEvent | CustomEvent) => {
+      if ((ev as StorageEvent).key && (ev as StorageEvent).key !== key) {
         return
       }
       setStoredValue(readValue())
